moved non Doctrine class to according ns
