+++ src/Akeneo/Pim/Enrichment/Bundle/EventSubscriber/Product/OnDelete/DeleteCompletenessSubscriber.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace Akeneo\Pim\Enrichment\Bundle\EventSubscriber\Product\OnDelete;
 
-use Akeneo\Pim\Enrichment\Bundle\Doctrine\Common\Remover\CompletenessRemover;
+use Akeneo\Pim\Enrichment\Bundle\Storage\Sql\Product\CompletenessRemover;
 use Akeneo\Pim\Enrichment\Component\Product\Model\ProductInterface;
 use Akeneo\Tool\Component\StorageUtils\Event\RemoveEvent;
 use Akeneo\Tool\Component\StorageUtils\StorageEvents;
