+++ src/Akeneo/Pim/Enrichment/Bundle/Resources/config/removers.yml
@@ -3,7 +3,7 @@ parameters:
 
 services:
     pim_catalog.completeness.remover:
-        class: 'Akeneo\Pim\Enrichment\Bundle\Doctrine\Common\Remover\CompletenessRemover'
+        class: 'Akeneo\Pim\Enrichment\Bundle\Storage\Sql\Product\CompletenessRemover'
         arguments:
             - '@database_connection'
